test: move remaining Node-only guards into *.node.test.ts files
Converts the last 7 inline isNode/isEdge guard sites (http.test.ts,
config.test.ts, assets.test.ts, projects.test.ts, listeners.test.ts) into
guard-free *.node.test.ts siblings, matching the convention http.node.test.ts
already established. Renames proxy.test.ts and resolveEventSourceFetch.test.ts
to *.node.test.ts and drops their now-redundant describe.skipIf guards, since
nonNodeExclude already keeps *.node.test.ts out of non-Node runs.

Pure move: no assertions changed, no tests added or removed. The non-Node
runs no longer collect-then-skip these; they are excluded outright by
filename. The only skip left in those runs is the pre-existing,
environment-agnostic test.skip in exports.test.ts.
