+++ test/client/projects.test.ts
@@ -4,7 +4,7 @@ import {firstValueFrom} from 'rxjs'
 import {describe, expect, test, vi} from 'vitest'
 
 import {getActiveFetch, getActiveMock} from '../helpers/mockFetch'
-import {apiHost, createClient, getClient, isNode} from './helpers'
+import {apiHost, createClient, getClient} from './helpers'
 
 describe('PROJECTS', () => {
   test('can request list of projects', async () => {
@@ -279,30 +279,6 @@ describe('PROJECTS', () => {
     expect(error).toBeInstanceOf(TypeError)
     expect(error).toMatchObject({message: 'Request options must include a `url`'})
   })
-
-  test.runIf(isNode)(
-    'the raw requester export leaves no listeners on a reused caller signal',
-    async () => {
-      getActiveMock()
-        .scope(`https://${apiHost}`)
-        .on('GET', '/v1/ping')
-        .respondPersist({status: 200, body: {pong: true}})
-
-      const controller = new AbortController()
-      for (let i = 0; i < 3; i++) {
-        await firstValueFrom(
-          requester({
-            url: `https://${apiHost}/v1/ping`,
-            signal: controller.signal,
-            fetch: getActiveFetch(),
-          }),
-        )
-      }
-
-      const {default: nodeEvents} = await import('node:events')
-      expect(nodeEvents.getEventListeners(controller.signal, 'abort')).toHaveLength(0)
-    },
-  )
 
   test('the raw requester export is lazy and cold', async () => {
     getActiveMock()
